refactor: more structure now that we're growing
