+++ aec/util/configure.py
