+++ tests/test_sqs.py
@@ -2,7 +2,7 @@ import boto3
 import pytest
 from moto import mock_sqs
 
-from aec.sqs import drain
+from aec.command.sqs import drain
 
 
 @pytest.fixture
